Base off of oraclelinux:7-slim
This dockerfile was written with RPM in mind which was brought in via the openjdk image
which is based on the oraclelinux:7-slim image. We could use the openjdk:8 image instead,
but that is debian-based which would require rewriting the yum invocation with apt. Opting
for least change, it seemed best to use oraclelinux:7-slim.